requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
jinja2==3.1.4
python-dotenv==1.0.1
pytest>=9.0
